+++ Weather_app/exceptions.py
